Changed data type of array
Changed the board array from string type to char type. The for loop now initializes all the elements of that array to a space (ASCII code = 32). Now in the printBoard method, it prints the square brackets ([,]) before and after the board piece. Finally, deleted the for loop that was contained in comments.