User input and board interaction
Added the ability for a user to input a location on the board to place a chip. Also added the ability for the users input to interact with the game board. The chip will fall down in the correct place for the user's turn. The chip will also pile up on any other chip that's beneath it (the program detects that there is a chip and places the new chip on top of it). PROBLEM, it only does one colour (for one user) and is not in its own method. Will be added later.