User input method
Created a user input method that will take in user's input, verify whether or not it's a valid input, then return that value for the rest of the program to interpret. Uses try and catch statements to catch any non integer entries and an if statement to verify whether or not the value can be interpreted by the rest of the program (if the value is within 1-7). The method automatically subtracts 1 from the entry so that it doesn't cause array out of bounds errors.