Two Player and Board method
Added two player capability in the program. Through using the modulo function and if statements, the program will cycle through each players turn.  As well, I placed the code that puts has board interaction in its own method. The 2D board array had to be made public in order to do this, however it makes the code more efficient through being able to call the that method for each players turn.