Winning is possible
Now win conditions have been added to the game so that someone can win. These lines of code are inside the winChecker method, which returns a boolean value based on what it reads from the board. It does this by looking at each piece that is placed then assess' whether or not it has the appropriate number of chips proceeding it to qualify as a win. Right now it only works for the horizontal and vertical directions, further updates will bring the diagonal function. Overall a big update that changes the program quite a bit.