Win conditions complete
Finally added the win conditions for both diagonal directions (lower left to upper right and lower right to upper left). It has been tested and works for most conditions, however it will be tested more heavily. Also made a method for setting up the board. The method takes all the elements of the board array and sets them as a white space to initialize and clean the board after a game has finished. Also added a some statements that say which player has one after the winChecker method has run and checked if there is a winner. Finally, added a boolean variable called gameOver that will set to true when someone has won to end the loop.