Game is basically complete
Very productive update as it basically completes the game. First off, added a functionality that detects whether or not a column is full on the board and tells the user to re-enter their column value if they entered one that is full. Second, added a replay function that asks the user whether or not they would like to play the game again. This is done with the help of a return method that returns a boolean variable based on their response. Game shall be tested and errors that are seen shall be fixed.