Fixed the input bug
Fixed the bug where the program would allow the input of a number followed by a letter. (ex. 6e would be read as 6 and not denied) This was achieved by turning the initial entry by the user into a string, then parsing it afterwards into another variable of type int. If it couldn't parse the entry, brought up an error, which would be caught by the catch statement.