Main menu and AI preps
This update adds a main menu with a navigation system to let you access the options that are available. Also moved some loops to conform only to the player vs player game mode. Now when a player vs player game finishes, asks the user whether or not they would like to continue in that game mode, if they choose not too, returns to main menu. Finally, added preparations for the implementation of an AI opponent. Added a section in the main method where the AI versus mode code will go and be organized.