Computer Opponent
Added the option of playing against a computer opponent. The AI is formed by a random number generator that takes a number between 1 and 7 for its move. It's not a difficult opponent. Well... It's no opponent at all really. However, if you let it, it is able to win the game. The boardInteraction method was changed to implement the AI opponent. Instead of using the userInput method inside the boardInteraction method, it uses it as a parameter.  The method will now take in a move parameter, which is where it'll try and place the piece based off of it.